* unit visibility test has been separated from map rendering code and is no more based on yardmap (fixes #194)

+++ trunk/src/ta3d/src/ta3dbase.h
@@ -40,7 +40,7 @@ namespace TA3D
 
 
 	extern const float	player_color[30];
-	extern byte		player_color_map[10];
+	extern unsigned int	player_color_map[10];
 
 	extern int expected_players;
